Fix bug in `write_dot_hsv/3`
Value should be passed directly to `dot_write/4` as `Illumination`. Previously
is was being calculated as a hardware value (between 0-31) before being passed
to `dot_write`, which is where this calculation needs to take place, thus
reducing the maximum brightness with `write_dot_hsv` to aproximatly 1/3 of the
possible brightness.

Signed-off-by: Winford <winford@object.stream>