Remove color order hard coding from write_dot
Now uses `COLOR_ORDER` defined in `ap102.hrl`, other devices can be added with
a different color order defined.  In the future it may be more convientient to
have the color order configured by the user during init, or at least be able to
change the order in a user config setting that would over-ride the default
defined for the hardware.

Signed-off-by: Winford <winford@object.stream>