Fixup one_dot example
Speeds up and reduces the numer of cycley of the hsv sweep.
Lenthens the maximum random time to just over a second (1023 ms), by increasing
the number of random bytes used to 2, and dividing by 64. I tested first just
using 10 of the 16 bits, but this seemed to yeild a much more "average" kind
of duration, but by using all 16 bits and dividing the results seem to yeild a
slightly higher number of short times, giving a better flashy, or twinkling
effect.

Signed-off-by: Winford <winford@object.stream>